requirements.txt
discord.py
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-compute